test: cover registry manager
