+++ src/index.test.ts
@@ -4,7 +4,6 @@ import {
   setRegistry,
   resetRegistry,
   getCommands,
-  createRegistryManager,
 } from './index';
 import { createRegistry } from './registry';
 
@@ -48,15 +47,5 @@ describe('registry', () => {
   it('getCommands uses provided registry', () => {
     const custom = createRegistry({ commands: { foo: () => {} } });
     expect(getCommands(custom).foo).toBeDefined();
-  });
-});
-
-describe('createRegistryManager', () => {
-  it('creates isolated registries', () => {
-    const mgr1 = createRegistryManager();
-    const mgr2 = createRegistryManager();
-    mgr1.initRegistry({ commands: { foo: () => {} } });
-    expect(mgr1.getCommands().foo).toBeDefined();
-    expect(mgr2.getCommands().foo).toBeUndefined();
   });
 });
