+++ src/services/__tests__/lights.test.ts
@@ -1,19 +1,13 @@
-import { describe, it, expect, vi, beforeEach } from 'vitest';
+import { vi } from 'vitest';
 import type { LightCycle } from '../../domain/types';
 import { createLight, getUpcomingPhase, upsertPhases } from '../lights';
 
 const singleLight = vi.fn();
-interface LightInsert {
-  select: () => { single: typeof singleLight };
-}
-const insert = vi.fn<[unknown], LightInsert>();
+const insert = vi.fn();
 
 const singleCycle = vi.fn();
-interface CycleSelect {
-  eq: () => { order: () => { limit: () => { single: typeof singleCycle } } };
-}
 const upsert = vi.fn();
-const selectCycle = vi.fn<[], CycleSelect>();
+const selectCycle = vi.fn();
 
 vi.mock('../../lib/supabase', () => ({
   supabase: {
